requirements.txt
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.2
docutils==0.14
enum34==1.1.6
functools32==3.2.3.post2
idna==2.8
imgcat==0.3.0
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.4.0
nbformat==4.4.0
numpy==1.16.0
Pillow==6.0.0
plotly==3.5.0
psutil==5.5.0
pytz==2018.9
requests==2.21.0
retrying==1.3.3
six==1.12.0
statistics==1.0.3.5
traitlets==4.3.2
urllib3>=1.24.2
